update vue with changes due to changes to replyvp API (#32)
* fix: restore catch block

* docs(README): update api info

* refactor:  get credentials from associative array

* refactor: get credentials from associative array with uniqid

* test: add test to cover user registration, login and token gen

* fix: correct assign generated credentials

* test: add test for ticket creation

* refactor: change to return named parameters

* refactor: add status and created_at to response body

* fix: correct getStatus return type

* fix: allow the user to retrieve a created ticket by id

* test: add test to cover ticket fetching

* fix: correct return type from string to int

* test: add test to cover posting a new message to a ticket

* test: add test to cover fetching all ticket messages

* refactor: remove unused addNewSupportReply function

* feat: define api endpoints

* refactor: assign new endpoint to register

* refactor: change endpoint to get tickets

* refactor: remove id parameter from getTickets

* refactor: remove id from getTickets

* refactor : change endpoint to get pending tickets

* refactor: remove id parameter from getTickets function

* refactor: change to get all logged user tickets

* refactor: change to send bearer token

* refactor: remove id parameter

* refactor: remove id parameter

* refactor: change endpoint to post messages

* refactor: remove id parameter from addNewMessage

* refactor: change to send bearer token

* fix: assign correct parameter

* fix: assign correct api endpoint

* refactor: assign endpoints

* refactor: change to remove id

* refactor: remove id parameter

* refactor: change to send bearer token

* refactor: change endpoint

* refactor: change endpoint

* refactor: change endpoint to terminate account

* refactor: change endpoint to update ticket status

* refactor: change endpoint to fetch all users

* refactor: remove unused endpoint

* refactor: change to return full response

* refactor: remove unused imports

* refactor: remove unused code

* refactor: check token existence

+++ client/app/src/services/adminService.js
@@ -1,14 +1,20 @@
 import axios from "axios";
-
-const ADMIN_API_URL = "http://localhost:8080/api/users";
+import { API_ENDPOINTS } from "./api";
+import authService from "@/services/authService";
 
 const adminService = {
   
-  async getAllUsers(token) {
+  async getAllUsers() {
+    const token = authService.getUserToken();
+    
+    if (!token) {
+      return false;
+    }
+
     try {
 
       const response = await axios.get(
-        `${ADMIN_API_URL}/all_users.php`,
+        API_ENDPOINTS.TICKETS.USERS,
         {
           headers: {
             Authorization: `Bearer ${token}`,
@@ -25,16 +31,22 @@ const adminService = {
     console.log(error);
   }
   },
+  async updateUserRoles(userId, newRoles) {
+    const token = authService.getUserToken();
+    
+    if (!token) {
+      return false;
+    }
 
-  async updateUserRoles(userId, newRoles, token) {
     try {
       const response = await axios.put(
-        `${ADMIN_API_URL}/update_roles.php`,
-        { userId, roles: newRoles },
-        {
+        API_ENDPOINTS.USERS.ROLES(userId), {
           headers: {
             Authorization: `Bearer ${token}`,
           },
+        },
+        { 
+          roles: newRoles 
         }
       );
       if (!response.data.success) {
@@ -44,10 +56,16 @@ const adminService = {
       console.log(error);
     }
   },
-  async deleteUserAsAdmin(userId, token) {
+  async deleteUserAsAdmin(userId) {
+    const token = authService.getUserToken();
+    
+    if (!token) {
+      return false;
+    }
+    
     try {
       const response = await axios.delete(
-       `${ADMIN_API_URL}/delete_user.php?userId=${userId}`,
+       API_ENDPOINTS.USERS.BY_ID(userId),
         {
           headers: {
             Authorization: `Bearer ${token}`,
@@ -56,16 +74,17 @@ const adminService = {
       );
 
       if (response.data.success) {
-        this.users = this.users.filter((user) => user.id !== userId);
         alert("Usuário removido com sucesso.");
+        return response;
       } else {
         alert("Erro ao remover o usuário.");
+        return response;
       }
     } catch (error) {
       console.log(error);
+      throw error;
     }
-  }
-    
+  },
 };
 
 export default adminService;
